+++ core/src/com/teixeirarios/mad/lib/domain/entities/enemy/EnemyStatus.java
@@ -2,9 +2,9 @@ package com.teixeirarios.mad.lib.domain.entities.enemy;
 
 public class EnemyStatus {
 
-    public int maxHealth, currentHealth, damage;
+    public float maxHealth, currentHealth, damage;
 
-    public EnemyStatus(int maxHealth, int damage) {
+    public EnemyStatus(int maxHealth, float damage) {
         this.maxHealth = maxHealth;
         this.currentHealth = maxHealth;
         this.damage = damage;
